# docs/api-spec.yaml
openapi: "3.0.3"
info:
  title: "AgentX Platform API"
  description: "AgentX 电商 AI 工作助手 API 接口定义"
  version: "1.0.0"
  contact:
    name: "AgentX Team"
servers:
  - url: "http://localhost:8000"
    description: "本地开发环境"
  - url: "https://api.agentx.example.com"
    description: "生产环境"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 10000
          description: "用户消息"
          example: "帮我找美妆达人"
        company_id:
          type: string
          description: "公司 ID（租户隔离）"
        agent_name:
          type: string
          description: "指定 Agent 名称（可选，不指定时 Master 自动路由）"
        session_id:
          type: string
          description: "会话 ID（多轮对话上下文追踪）"
        mode:
          type: string
          enum: [workflow, agent, engine_first, hybrid]
          description: "工作模式（可选，默认 LLM 自动判断）"

    # ---- SSE 事件类型 ----
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          description: "对话标题（自动生成或用户设置）"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
        last_message:
          type: string
          description: "最后一条消息摘要"
    ConversationDetail:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
    Message:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, master, system]
        content:
          type: string
        content_type:
          type: string
          enum: [text, kol_list, analysis_report, script, logistics, plan]
        metadata:
          type: object
          description: "结构化元数据（达人列表 JSON、报告数据等）"
        references:
          type: array
          items:
            $ref: "#/components/schemas/Reference"
        created_at:
          type: string
          format: date-time
    Reference:
      type: object
      properties:
        source_file:
          type: string
        source_page:
          type: integer
        score:
          type: number
        content:
          type: string
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200

    # ---- 达人 ----
    KolProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "李佳琦Austin"
        platform:
          type: string
          enum: [douyin, xiaohongshu, kuaishou, bilibili, weibo]
        followers:
          type: integer
          description: "粉丝数"
          example: 48500000
        engagement_rate:
          type: number
          description: "互动率 (百分比)"
          example: 3.5
        category:
          type: string
          example: "美妆"
        avg_views:
          type: integer
          description: "平均播放量"
        avg_likes:
          type: integer
          description: "平均点赞数"
        avg_comments:
          type: integer
          description: "平均评论数"
        price_range:
          type: string
          description: "报价区间"
          example: "8000-15000"
        location:
          type: string
          description: "所在地"
        verified:
          type: boolean
        bio:
          type: string
          description: "个人简介"
        avatar_url:
          type: string
        created_at:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/chat:
    post:
      tags: [chat]
      summary: "流式聊天（SSE）"
      description: |
        核心聊天接口，使用 Server-Sent Events (SSE) 实现流式响应。

        SSE 事件类型:
        - `thinking`: AI 正在思考
        - `intent`: 意图识别结果
        - `plan`: 执行计划
        - `sources`: RAG 引用来源
        - `tool_result`: 工具执行结果
        - `content`: 增量文本内容
        - `reflection`: 质量审查结果
        - `retry`: 重试状态
        - `done`: 对话完成
        - `error`: 错误信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: "SSE 事件流"
          content:
            text/plain:
              schema:
                type: string
                description: "SSE 格式事件流"
        "401":
          description: "未登录"
...
  /api/conversations:
    get:
      tags: [conversations]
      summary: "获取对话列表"
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: "对话列表"
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"

    post:
      tags: [conversations]
      summary: "创建新对话"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: "创建成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /api/conversations/{conversation_id}:
    get:
      tags: [conversations]
      summary: "获取对话详情（含消息列表）"
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "对话详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "404":
          description: "对话不存在"

    delete:
      tags: [conversations]
      summary: "删除对话"
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: "删除成功"
        "404":
          description: "对话不存在"

  # ============================================================
  # 达人搜索
  # ============================================================
...
  /api/kol/{kol_id}:
    get:
      tags: [kol]
      summary: "达人详情"
      parameters:
        - name: kol_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "达人详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KolProfile"
        "404":
          description: "达人不存在"
...